Show Event Details

See merge request krh951/cmpt370-9mare!65

+++ app/src/main/java/com/example/cmpt370_9mare/ui/event/CreateEventFragment.kt
@@ -318,11 +318,6 @@ class CreateEventFragment : Fragment() {
         }
     }
 
-    fun onSelectRepeat() {
-        val action = CreateEventFragmentDirections.actionCreateEventFragmentToNewEventFragment()
-        findNavController().navigate(action)
-    }
-
     fun showDatePicker() {
         val date = binding.inputDate.text.toString()
         DatePickerFragment(date).show(childFragmentManager, DatePickerFragment.DATE_PICKER)
@@ -430,12 +425,23 @@ class CreateEventFragment : Fragment() {
         binding.submitCreateEvent.visibility = if (boolean) View.VISIBLE else View.INVISIBLE
     }
 
+    private fun showDeleteConfirmationDialog() {
+        val builder = AlertDialog.Builder(this.context)
+        builder.setMessage("Delete the current event?")
+            .setCancelable(false)
+            .setPositiveButton("Confirm") { _, _ ->
+                scheduleEventShareViewModel.deleteItem(currentEvent)
+                findNavController().navigateUp()
+            }
+            .setNegativeButton("Cancel") { dialog, _ -> dialog.cancel() }
+            .show()
+    }
+
     /**
      * Delete Event
      */
     fun deleteEvent() {
-        scheduleEventShareViewModel.deleteItem(currentEvent)
-        findNavController().navigateUp()
+        showDeleteConfirmationDialog()
     }
